refactor denormalization, registry close and registry into separate files
